fix(retro): show read-only closed summary for retros marked closed in DB
Closed retros (status='closed') were rendering the name-prompt join
lobby to anyone with the link. The page is now a server component that
queries the DB by roomCode first. If status='closed', it renders
ClosedRetroSummary (groups ranked by votes, cards by category, action
items per group plus a general section) with no interactive controls and
no PartyKit connection. If the DB query fails or the retro is still
active, it falls through to the existing RetroRoomClient unchanged.
